requirements.txt
Django
django-fcm
ipython
psycopg2-binary
django-oauth-toolkit
gunicorn
